+++ besca/st/__init__.py
@@ -1,5 +1,6 @@
 from ._FAIR_export import export_cp10k, export_regressedOut, export_clr, export_clustering, export_metadata, export_rank, export_celltype
-from ._wrapper_funcs import setup, setup_citeseq, read_matrix, filtering_cells_genes_min, filtering_mito_genes_max, per_cell_normalize, clr_normalize, dsb_normalize, highly_variable_genes, regress_out, batch_correction, pca_neighbors_umap, clustering, additional_labeling, celltype_labeling
+from ._wrapper_funcs import setup, setup_citeseq, read_matrix, filtering_cells_genes_min, filtering_mito_genes_max, per_cell_normalize, clr_normalize, highly_variable_genes, regress_out, batch_correction, pca_neighbors_umap, clustering, additional_labeling, celltype_labeling
+from ._wrapper_Rfuncs import dsb_normalize
 from ._setup_funcs import create_button, create_popup
 from ._qc_report import write_qc
